requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.8
django-import-export==2.6.1
django-simpleui==2021.8.17
et-xmlfile==1.1.0
idna==3.3
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
Pillow==8.3.2
psutil==5.8.0
pyasn1==0.4.8
pytz==2021.3
PyYAML==6.0
requests==2.26.0
rsa==4.7.2
simplepro==3.1.2
sqlparse==0.4.2
tablib==3.0.0
urllib3==1.26.7
xlrd==2.0.1
xlwt==1.3.0
